requirements.txt
streamlit
pandas
numpy
scipy
matplotlib
plotly
seaborn
yfinance
requests
fredapi
statsmodels
prophet
tensorflow
scikit-learn
python-dotenv
lxml
html5lib
webdriver-manager
